==> 2_prepare_inputs_to_DT.ipynb <==
import os
import pandas as pd

from finrl.main import check_and_make_directories                               
from finrl.config import INDICATORS, TRAINED_MODEL_DIR, RESULTS_DIR

# "./" will be added in front of each directory
check_and_make_directories([TRAINED_MODEL_DIR])


### Step1: Load the data

train = pd.read_csv('data/train.csv')

train = train.set_index(train.columns[0])
train.index.names = ['']


train.head()


train.shape


# train.columns
train.tic.unique(), INDICATORS


### Step 2: Setup the environment

state_space = 1 (remaining balance in the account) + 2stock_dimension (prices of 30 stocks and the share holdings of the 30 stocks, so totally 2stock_dimension) + len(INDICATORS)*stock_dimension

stock_dimension = len(train.tic.unique())
state_space = 1 + 2*stock_dimension + len(INDICATORS)*stock_dimension
print(f"Stock Dimension: {stock_dimension}, State Space: {state_space}")


Action: The action space describes the allowed actions that the agent interacts with the environment. Normally, a ∈ A includes three actions: a ∈ {−1, 0, 1}, where −1, 0, 1 represent selling, holding, and buying one stock. Also, an action can be carried upon multiple shares. We use an action space {−k, ..., −1, 0, 1, ..., k}, where k denotes the number of shares. For example, "Buy 10 shares of AAPL" or "Sell 10 shares of AAPL" are 10 or −10, respectively

action is a scalar in a continuous space instead of a discrete space like {-1, 0, 1}. Hence, we just need one action for each stock, and that action is from a continuous space. Thus the action is always a vector with stock dimension.

from finrl.meta.env_stock_trading.env_stocktrading import StockTradingEnv

buy_cost_list = sell_cost_list = [0.001] * stock_dimension
num_stock_shares = [0] * stock_dimension

env_kwargs = {
    "hmax": 100,
    "initial_amount": 1000000,
    "num_stock_shares": num_stock_shares,
    "buy_cost_pct": buy_cost_list,
    "sell_cost_pct": sell_cost_list,
    "state_space": state_space,
    "stock_dim": stock_dimension,
    "tech_indicator_list": INDICATORS,
    "action_space": stock_dimension,
    "reward_scaling": 1e-4
}

e_train_gym = StockTradingEnv(df = train, **env_kwargs)
env_train, _ = e_train_gym.get_sb_env()


len(e_train_gym.df.index.unique()) - 1


e_train_gym.df.tic.count()


### Step 3: Initially train using a PPO Model

from stable_baselines3 import PPO
from finrl.agents.stablebaselines3.models import DRLAgent
from stable_baselines3.common.logger import configure

agent = DRLAgent(env = env_train)
model_ppo = agent.get_model('ppo')

# set up logger
tmp_path = RESULTS_DIR + '/ppo'
new_logger_ppo = configure(tmp_path, ["stdout", "csv", "tensorboard"])
# Set new logger
model_ppo.set_logger(new_logger_ppo)

trained_ppo = agent.train_model(model=model_ppo,
                             tb_log_name='ppo',
                             total_timesteps=50000)


trained_ppo.save(TRAINED_MODEL_DIR + "/agent_ppo")


# Load the saved model

model = PPO.load("trained_models/agent_ppo.zip")


### Step 4: Use PPO Model to generate trajectories

import numpy as np

"""make a prediction and get results"""
env_train, obs = e_train_gym.get_sb_env()

ds = []
states = []
feature = {}

s, a, r, d = [], [], [], []

env_train.reset()
#max_steps = len(e_train_gym.df.index.unique()) - 1
max_steps = e_train_gym.df.tic.count() - 1

for i in range(1, max_steps, 1):

    action, _states = model.predict(obs, deterministic=True)
    s.extend(obs)
    a.extend(action)

    obs, rewards, dones, info = env_train.step(action)
    r.extend(rewards)
    d.append(dones[0])

    states.extend(obs)

    if (i % 100 == 0):
        
        feature['observations'] = s
        feature['actions'] = a
        feature['rewards'] = r
        feature['dones'] = d
        
        ds.append(feature)
        feature = {}
        s, a, r, d = [], [], [], []


states = np.vstack(states)
state_mean, state_std = np.mean(states, axis=0), np.std(states, axis=0) + 1e-6


state_mean[:5], state_std[:5], state_mean.shape


len_ds = len(ds)

state_mean = np.pad(state_mean, (0, (len_ds-state_space)))
state_std = np.pad(state_std, (0, (len_ds-state_space)))


state_mean, len(state_mean)


len(ds), len(ds[0])


feature = ds[0]
len(feature["rewards"])


https://huggingface.co/docs/datasets/v1.11.0/loading_datasets.html#from-local-files

csv
- dataset = load_dataset(‘csv’, data_files=[‘my_file_1.csv’, ‘my_file_2.csv’])

json
- dataset = load_dataset(‘json’, data_files=’my_file.json’)

text
- dataset = load_dataset(‘text’, data_files={‘train’: [‘my_text_1.txt’, ‘my_text_2.txt’], ‘test’: ‘my_test_file.txt’})

python dict
- my_dict = {'id': [0, 1, 2], 'name': ['mary', 'bob', 'eve'], 'age': [24, 53, 19]}
- dataset = Dataset.from_dict(my_dict)

pandas dataframe
- df = pd.DataFrame({"a": [1, 2, 3]})
- dataset = Dataset.from_pandas(df)


### Step 5: Wrap the trajectory data using a dictionary

input_data = {}
input_data['train'] = ds
input_data['state_mean'] = state_mean
input_data['state_std'] = state_std


input_data.keys()


from datasets import Dataset

dataset = Dataset.from_dict(input_data)


### Step 6: Save the trajectory data to the local disk

https://huggingface.co/docs/datasets/v1.10.2/processing.html

dataset.save_to_disk("data/dataset/")


from datasets import load_from_disk

dataset = load_from_disk("data/dataset/")
